+++ sql/catalyst/src/main/scala/org/apache/spark/sql/internal/SQLConf.scala
@@ -1940,10 +1940,6 @@ object SQLConf {
       // 5 is the default table format version for RocksDB 6.20.3.
       .createWithDefault(5)
 
-  /**
-   * Note: this is defined in `RocksDBConf.FORMAT_VERSION`. These two places should be updated
-   * together.
-   */
   val STATE_STORE_ROCKSDB_CHANGE_CHECKPOINTING_ENABLED =
     buildConf("spark.sql.streaming.stateStore.rocksdb.enableChangelogCheckpointing")
       .internal()
