--- sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/state/Changelog.scala
@@ -1,153 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.sql.execution.streaming.state
-
-import java.io.{DataInputStream, DataOutputStream, FileNotFoundException, IOException}
-
-import scala.util.control.NonFatal
-
-import com.google.common.io.ByteStreams
-import org.apache.commons.io.IOUtils
-import org.apache.hadoop.fs.{FSError, Path}
-
-import org.apache.spark.internal.Logging
-import org.apache.spark.io.CompressionCodec
-import org.apache.spark.sql.execution.streaming.CheckpointFileManager
-import org.apache.spark.sql.execution.streaming.CheckpointFileManager.CancellableFSDataOutputStream
-
-
-class ChangelogReader(fm: CheckpointFileManager, fileToRead: Path,
-                      compressionCodec: CompressionCodec)
-  extends Iterator[ByteArrayPair] with Logging {
-  private def decompressStream(inputStream: DataInputStream): DataInputStream = {
-    val compressed = compressionCodec.compressedInputStream(inputStream)
-    new DataInputStream(compressed)
-  }
-  private val sourceStream = try {
-    fm.open(fileToRead)
-  } catch {
-    case f: FileNotFoundException =>
-      throw new IllegalStateException(
-        s"Error reading streaming state file of $this: $fileToRead does not exist. " +
-          s"If the stream job is restarted with a new or updated state operation, please" +
-          s" create a new checkpoint location or clear the existing checkpoint location.", f)
-  }
-  private val input: DataInputStream = decompressStream(sourceStream)
-  private var byteArrayPair: ByteArrayPair = null
-  private var eof = false
-
-  override def hasNext: Boolean = {
-    maybeReadNext()
-    byteArrayPair != null
-  }
-
-  override def next(): ByteArrayPair = {
-    maybeReadNext()
-    val nextByteArrayPair = byteArrayPair
-    byteArrayPair = null
-    nextByteArrayPair
-  }
-
-  def close(): Unit = { if (input != null) input.close() }
-
-  private def maybeReadNext(): Unit = {
-    if (!eof && byteArrayPair == null) {
-      val keySize = input.readInt()
-      if (keySize == -1) {
-        eof = true
-      } else if (keySize < 0) {
-        throw new IOException(
-          s"Error reading streaming state file $fileToRead of $this: key size cannot be $keySize")
-      } else {
-        val keyBuffer = new Array[Byte](keySize)
-        ByteStreams.readFully(input, keyBuffer, 0, keySize)
-        val valueSize = input.readInt()
-        if (valueSize < 0) {
-          byteArrayPair = new ByteArrayPair(keyBuffer, null)
-        } else {
-          val valueBuffer = new Array[Byte](valueSize)
-          ByteStreams.readFully(input, valueBuffer, 0, valueSize)
-          byteArrayPair = new ByteArrayPair(keyBuffer, valueBuffer)
-        }
-      }
-    }
-  }
-}
-
-class ChangelogWriter(fm: CheckpointFileManager, file: Path,
-                      compressionCodec: CompressionCodec) extends Logging {
-  private def compressStream(outputStream: DataOutputStream): DataOutputStream = {
-    val compressed = compressionCodec.compressedOutputStream(outputStream)
-    new DataOutputStream(compressed)
-  }
-
-  private var backingFileStream: CancellableFSDataOutputStream =
-    fm.createAtomic(file, overwriteIfPossible = true)
-  private var compressedStream: DataOutputStream = compressStream(backingFileStream)
-
-  def put(key: Array[Byte], value: Array[Byte]): Unit = {
-    assert(compressedStream != null)
-    compressedStream.writeInt(key.size)
-    compressedStream.write(key)
-    compressedStream.writeInt(value.size)
-    compressedStream.write(value)
-  }
-
-  def delete(key: Array[Byte]): Unit = {
-    assert(compressedStream != null)
-    compressedStream.writeInt(key.size)
-    compressedStream.write(key)
-    compressedStream.writeInt(-1)
-  }
-
-  def abort(): Unit = {
-    try {
-      if (backingFileStream != null) backingFileStream.cancel()
-      if (compressedStream != null) IOUtils.closeQuietly(compressedStream)
-    } catch {
-      // Closing the compressedStream causes the stream to write/flush flush data into the
-      // rawStream. Since the rawStream is already closed, there may be errors.
-      // Usually its an IOException. However, Hadoop's RawLocalFileSystem wraps
-      // IOException into FSError.
-      case e: FSError if e.getCause.isInstanceOf[IOException] =>
-
-      // SPARK-42668 - Catch and log any other exception thrown while trying to cancel
-      // raw stream or close compressed stream.
-      case NonFatal(ex) =>
-        logInfo(s"Failed to cancel delta file for provider " +
-          s"with exception=$ex")
-    } finally {
-      backingFileStream = null
-      compressedStream = null
-    }
-  }
-
-  def commit(): Unit = {
-    try {
-      compressedStream.writeInt(-1)
-      compressedStream.close()
-    } catch {
-      case e: Throwable =>
-        abort()
-        throw e
-    } finally {
-      backingFileStream = null
-      compressedStream = null
-    }
-  }
-}
